[ray_client]: Move from experimental to util
