+++ python/ray/tests/BUILD
@@ -75,20 +75,20 @@ py_test_module_list(
   files = [
     "test_actor_pool.py",
     "test_args.py",
-    "test_asyncio.py",
     "test_asyncio_cluster.py",
+    "test_asyncio.py",
     "test_autoscaler.py",
     "test_autoscaler_yaml.py",
-    "test_component_failures.py",
+    "test_client_metadata.py",
+    "test_client.py",
+    "test_client_references.py",
+    "test_client_terminate.py",
     "test_command_runner.py",
+    "test_component_failures.py",
     "test_coordinator_server.py",
-    "test_dask_scheduler.py",
     "test_dask_callback.py",
+    "test_dask_scheduler.py",
     "test_debug_tools.py",
-    "test_experimental_client.py",
-    "test_experimental_client_metadata.py",
-    "test_experimental_client_references.py",
-    "test_experimental_client_terminate.py",
     "test_job.py",
     "test_memstat.py",
     "test_metrics_agent.py",
