+++ python/ray/tests/test_advanced.py
@@ -17,7 +17,7 @@ from ray.test_utils import client_test_enabled
 from ray.test_utils import RayTestTimeoutException
 
 if client_test_enabled():
-    from ray.experimental.client import ray
+    from ray.util.client import ray
 else:
     import ray
