+++ python/ray/tests/test_client_terminate.py
@@ -1,5 +1,5 @@
 import pytest
-from ray.experimental.client.ray_client_helpers import ray_start_client_server
+from ray.util.client.ray_client_helpers import ray_start_client_server
 from ray.tests.client_test_utils import create_remote_signal_actor
 from ray.test_utils import wait_for_condition
 from ray.exceptions import TaskCancelledError
