+++ python/ray/util/client/server/dataservicer.py
@@ -7,7 +7,7 @@ import ray.core.generated.ray_client_pb2 as ray_client_pb2
 import ray.core.generated.ray_client_pb2_grpc as ray_client_pb2_grpc
 
 if TYPE_CHECKING:
-    from ray.experimental.client.server.server import RayletServicer
+    from ray.util.client.server.server import RayletServicer
 
 logger = logging.getLogger(__name__)
